requirements.txt
pycryptodomex
requests
bs4
